Minor remote server cleanup
